+++ bitcoind-rpc-test/src/test/scala/org/bitcoins/rpc/common/BitcoindVersionTest.scala
@@ -6,9 +6,9 @@ import org.bitcoins.testkit.util.BitcoindRpcTest
 class BitcoindVersionTest extends BitcoindRpcTest {
   behavior of "BitcoindVersion"
 
-  it should "return version 24" in {
-    val version = BitcoindVersion.fromNetworkVersion(240100)
-    assert(version.equals(BitcoindVersion.V24))
+  it should "return version 25" in {
+    val version = BitcoindVersion.fromNetworkVersion(250100)
+    assert(version.equals(BitcoindVersion.V25))
   }
 
 }
