+++ bitcoind-rpc-test/src/test/scala/org/bitcoins/rpc/common/BlockchainRpcTest.scala
@@ -291,4 +291,62 @@ class BlockchainRpcTest extends BitcoindFixturesCachedPairNewest {
       assert(!response2.asInstanceOf[ScanBlocksAbortResult].aborted)
     }
   }
+
+  it should "get a block filter given a block hash" in { case nodePair =>
+    val client = nodePair.node1
+    for {
+      blocks <- client.generate(1)
+      blockHashBE = blocks.head
+      blockFilter <- client.getBlockFilter(blockHashBE, FilterType.Basic)
+      block <- client.getBlockRaw(blockHashBE)
+      prevOuts = block.transactions
+        .filterNot(_.isCoinbase)
+        .flatMap(_.inputs.map(_.previousOutput))
+      fundingOutputs <- Future.traverse(prevOuts) { outpoint =>
+        client
+          .getTransaction(outpoint.txIdBE)
+          .map(_.hex.outputs(outpoint.idx))
+      }
+      prevFilter <- client.getBlockFilter(
+        block.blockHeader.previousBlockHashBE,
+        FilterType.Basic
+      )
+    } yield {
+      val pubKeys = fundingOutputs.map(_.scriptPubKey).toVector
+      val filter = BlockFilter(block, pubKeys)
+      assert(filter.hash == blockFilter.filter.hash)
+      assert(
+        blockFilter.header == filter
+          .getHeader(prevFilter.header.flip)
+          .hashBE
+      )
+    }
+  }
+  it should "be able to get blockchain info" in { case nodePair =>
+    val client = nodePair.node1
+    for {
+      info <- client.getBlockChainInfo
+      bestHash <- client.getBestBlockHash()
+    } yield {
+      assert(info.isInstanceOf[GetBlockChainInfoResultPostV23])
+      val postV23Info = info.asInstanceOf[GetBlockChainInfoResultPostV23]
+      assert(postV23Info.chain == RegTest)
+      assert(postV23Info.bestblockhash == bestHash)
+    }
+  }
+
+  it should "be able to get a block with verbose transactions" in {
+    case nodePair =>
+      val client = nodePair.node1
+      for {
+        blocks <- client.generate(2)
+        block <- client.getBlockWithTransactions(blocks(1))
+      } yield {
+        assert(block.hash == blocks(1))
+        assert(block.tx.length == 1)
+        val tx = block.tx.head
+        assert(tx.vout.head.n == 0)
+      }
+  }
+
 }
