+++ bitcoind-rpc-test/src/test/scala/org/bitcoins/rpc/common/UtilRpcTest.scala
@@ -2,7 +2,8 @@ package org.bitcoins.rpc.common
 
 import org.bitcoins.asyncutil.AsyncUtil
 import org.bitcoins.commons.jsonmodels.bitcoind.RpcOpts.AddressType
-import org.bitcoins.core.protocol.P2PKHAddress
+import org.bitcoins.core.protocol.{BitcoinAddress, P2PKHAddress}
+import org.bitcoins.core.protocol.script.descriptor.Descriptor
 import org.bitcoins.crypto.ECPrivateKey
 import org.bitcoins.rpc.client.common.BitcoindRpcClient
 import org.bitcoins.testkit.rpc.BitcoindFixturesCachedPairNewest
@@ -65,5 +66,56 @@ class UtilRpcTest extends BitcoindFixturesCachedPairNewest {
       assert(blockFilterIndexInfo.synced)
       assert(blockFilterIndexInfo.best_block_height == 101)
     }
+  }
+
+  it should "return active rpc commands" in { nodePair =>
+    val client = nodePair.node1
+    val generatedF =
+      client.getNewAddress.flatMap(addr => client.generateToAddress(100, addr))
+    val rpcinfoF =
+      generatedF.flatMap(_ => client.getRpcInfo())
+
+    rpcinfoF.map { result =>
+      assert(result.active_commands.length == 1)
+    }
+  }
+
+  it should "derive addresses from a descriptor" in { case nodePair =>
+    val client = nodePair.node1
+    val str0 =
+      "wpkh(tprv8ZgxMBicQKsPd7Uf69XL1XwhmjHopUGep8GuEiJDZmbQz6o58LninorQAfcKZWARbtRtfnLcJ5MQ2AtHcQJCCRUcMRvmDUjyEmNUWwx8UbK/1/1/0)#t6wfjs64"
+    val descriptor0 = Descriptor.fromString(str0)
+    assert(descriptor0.toString == str0)
+    val addresses0F =
+      client.deriveAddresses(descriptor0, None).map(_.addresses)
+    val expected0 =
+      Vector("bcrt1qjqmxmkpmxt80xz4y3746zgt0q3u3ferr34acd5").map(
+        BitcoinAddress.fromString
+      )
+    val assert0 = addresses0F.map { addresses =>
+      assert(addresses == expected0)
+    }
+
+    val str1 =
+      "wpkh(tprv8ZgxMBicQKsPd7Uf69XL1XwhmjHopUGep8GuEiJDZmbQz6o58LninorQAfcKZWARbtRtfnLcJ5MQ2AtHcQJCCRUcMRvmDUjyEmNUWwx8UbK/1/1/*)#kft60nuy"
+
+    val descriptor1 = Descriptor.fromString(str1)
+    assert(descriptor1.toString == str1)
+    val addresses1F =
+      client.deriveAddresses(descriptor1, Some(Vector(0, 2))).map(_.addresses)
+    val expected1 =
+      Vector(
+        "bcrt1qjqmxmkpmxt80xz4y3746zgt0q3u3ferr34acd5",
+        "bcrt1qhku5rq7jz8ulufe2y6fkcpnlvpsta7rq4442dy",
+        "bcrt1qpgptk2gvshyl0s9lqshsmx932l9ccsv265tvaq"
+      )
+        .map(BitcoinAddress.fromString)
+
+    val assert1 = assert0.flatMap(_ =>
+      addresses1F.map { addresses =>
+        assert(addresses == expected1)
+      })
+
+    assert1
   }
 }
