+++ bitcoind-rpc/src/main/scala/org/bitcoins/rpc/client/common/BitcoindRpcClient.scala
@@ -16,7 +16,6 @@ import org.bitcoins.core.wallet.fee.FeeUnit
 import org.bitcoins.crypto.{DoubleSha256DigestBE, StringFactory}
 import org.bitcoins.rpc.client.v18.V18AssortedRpc
 import org.bitcoins.rpc.client.v20.{V20AssortedRpc, V20MultisigRpc}
-import org.bitcoins.rpc.client.v24.BitcoindV24RpcClient
 import org.bitcoins.rpc.client.v25.BitcoindV25RpcClient
 import org.bitcoins.rpc.config._
 
@@ -344,7 +343,6 @@ object BitcoindRpcClient {
       system: ActorSystem
   ): BitcoindRpcClient = {
     val bitcoind = version match {
-      case BitcoindVersion.V24 => BitcoindV24RpcClient.withActorSystem(instance)
       case BitcoindVersion.V25 => BitcoindV25RpcClient.withActorSystem(instance)
       case BitcoindVersion.Unknown =>
         sys.error(
@@ -372,13 +370,9 @@ object BitcoindVersion
   val newest: BitcoindVersion = V25
 
   val standard: Vector[BitcoindVersion] =
-    Vector(V25, V24)
+    Vector(V25)
 
   val known: Vector[BitcoindVersion] = standard
-
-  case object V24 extends BitcoindVersion {
-    override def toString: String = "v24"
-  }
 
   case object V25 extends BitcoindVersion {
     override def toString: String = "v25"
@@ -402,7 +396,7 @@ object BitcoindVersion
   def fromNetworkVersion(int: Int): BitcoindVersion = {
     // need to translate the int 210100 (as an example) to a BitcoindVersion
     int.toString.substring(0, 2) match {
-      case "24" => V24
+      case "25" => V25
       case _ =>
         logger.warn(
           s"Unsupported Bitcoin Core version: $int. The latest supported version is ${BitcoindVersion.newest}"
